feat(gdn): route exact Cake checkpoint prefill

+++ test/registered/unit/layers/attention/test_gdn_prefill_backend_policy.py
@@ -177,7 +177,7 @@ class TestFlashInferGDNPrefillBackendPolicy(unittest.TestCase):
 
         torch.testing.assert_close(
             metadata.state_checkpoint_cu_starts,
-            torch.tensor([0, 0, 1, 2, 3, 5, 7]),
+            torch.tensor([0, 0, 1, 2, 3, 5, 7], dtype=torch.int32),
         )
         torch.testing.assert_close(metadata.track_ssm_h_src, torch.tensor([1, 2, 3, 6]))
         self.assertEqual(metadata.num_state_checkpoints, 7)
